corrr-package {corrr}	R Documentation

corrr: Correlations in R

Description

A tool for exploring correlations. It makes it possible to easily perform routine tasks when exploring correlation matrices such as ignoring the diagonal, focusing on the correlations of certain variables against others, or rearranging and visualizing the matrix in terms of the strength of the correlations.

Author(s)

Maintainer: Max Kuhn max@rstudio.com

Authors:

Simon Jackson

Jorge Cimentada

See Also

Useful links:

https://github.com/tidymodels/corrr

https://corrr.tidymodels.org

Report bugs at https://github.com/tidymodels/corrr/issues

[Package corrr version 0.4.5 Index]