dice {corrr}	R Documentation

Returns a correlation table with the selected fields only

Description

Returns a correlation table with the selected fields only

Usage

dice(x, ...)

Arguments

x
A correlation table, class cor_df

...
A list of variables in the correlation table

Examples

dice(correlate(mtcars), mpg, wt, am)

[Package corrr version 0.4.5 Index]