retract {corrr}	R Documentation

Creates a data frame from a stretched correlation table

Description

retract does the opposite of what stretch does

Usage

retract(.data, x, y, val)

Arguments

.data
A data.frame or tibble containing at least three variables: x, y and the value

x
The name of the column to use from .data as x

y
The name of the column to use from .data as y

val
The name of the column to use from .data to use as the value

Examples

x <- correlate(mtcars)
xs <- stretch(x)
retract(xs)

[Package corrr version 0.4.5 Index]